Add assets directory to http file server in main.go
Previously, there was no proper organizational system for serving files on the http server from the main.go file. Now, the server is programmed to not only serve files from the root directory, but also the assets directory. This increases file accessibility and streamlines file paths under the /assets directory.